requirements.txt
albumentations==0.5.2
einops==0.3.2
huggingface-hub==0.2.1
imageio==2.9.0
imagesize==1.3.0
imgaug==0.4.0
ipdb==0.13.9
ipython-genutils==0.2.0
ipywidgets==7.6.5
matplotlib==3.3.3
nltk==3.5
opencv-python==4.5.1.48
opencv-python-headless==4.5.3.56
pandas==1.2.0
numpy==1.23.4
pathtools==0.1.2
Pillow==8.1.0
pydicom==2.1.2
pytorch-lightning==1.5.10
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.2
segmentation-models-pytorch==0.2.0
sklearn==0.0
timm==0.4.12
tokenizers==0.12.1
torchmetrics==0.8.2
tqdm==4.56.0
transformers==4.18.0
wandb==0.12.17
